refactor(runtime): Move add-peer HTML helpers
Re-home the peer trust and visibility add-peer fragments into the neutral runtime.peers.html package to remove the remaining direct runtime.alerts -> clients.http dependency.

Update both consumers, move the matching tests, and keep the focused verification for the moved helper seam passing.
