+++ src/main/java/network/crypta/clients/http/ConnectionsToadlet.java
@@ -14,8 +14,6 @@ import java.util.StringTokenizer;
 import network.crypta.client.FetchException;
 import network.crypta.client.HighLevelSimpleClient;
 import network.crypta.clients.fcp.AddPeer;
-import network.crypta.clients.http.complexhtmlnodes.PeerTrustInputForAddPeerBoxNode;
-import network.crypta.clients.http.complexhtmlnodes.PeerVisibilityInputForAddPeerBoxNode;
 import network.crypta.config.ConfigException;
 import network.crypta.keys.FreenetURI;
 import network.crypta.l10n.NodeL10n;
@@ -23,6 +21,8 @@ import network.crypta.node.DarknetPeerNode.FRIEND_TRUST;
 import network.crypta.node.DarknetPeerNode.FRIEND_VISIBILITY;
 import network.crypta.node.PeerNodeStatus;
 import network.crypta.node.Version;
+import network.crypta.runtime.peers.html.PeerTrustInputForAddPeerBoxNode;
+import network.crypta.runtime.peers.html.PeerVisibilityInputForAddPeerBoxNode;
 import network.crypta.runtime.spi.ConfigPort;
 import network.crypta.runtime.spi.ConnectionsPageKind;
 import network.crypta.runtime.spi.ConnectionsPagePort;
