+++ src/main/java/network/crypta/runtime/peers/html/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Peer-related reusable HTML fragment builders shared by runtime and HTTP UI code.
+ *
+ * <p>This package contains small HTML node helpers that are not page controllers themselves but are
+ * reused across the runtime alert and HTTP toadlet layers.
+ */
+package network.crypta.runtime.peers.html;
